requirements.txt
Flask==3.1.2
flask-cors==6.0.1
python-dotenv==1.2.1

torch==2.9.0
torchvision==0.24.0
ultralytics==8.3.224
ultralytics-thop==2.0.18

opencv-python==4.12.0.88
numpy==2.2.6
pillow==12.0.0
requests==2.32.5

matplotlib==3.10.7
scipy==1.16.3
PyYAML==6.0.3
google-generativeai==0.8.5
